requirements.txt
maritalk
python-dotenv
tiktoken